revert calling cancel on monitor ( in subscribe )
we cannot cancel because we are surfacing the Monitor instance to the
user ( by passing it through the callback ). It was a good idea to
remove this ( from a GC point of view ) but syncify is using it. I will
take a look at syncify to see if I can avoid using the Monitor directly.